requirements.txt
Flask==2.0.2
moviepy==1.0.3
python-dotenv==0.19.2
waitress==2.0.0